refactor: remove two headers and replace global root into private folder
