+++ app/(private)/layout.tsx
@@ -1,17 +1,52 @@
 import { ErrorTooltip } from '@/_shared/components/ErrorTooltip';
 import { Header } from '@/_shared/components/Header';
 import { getUser } from '@/_shared/lib/getUser';
+import type { Metadata } from 'next';
 import { redirect } from 'next/navigation';
 
-export default async function Layout({ children }: { children: React.ReactNode }) {
+import { geistMono, geistSans } from '@/_shared/fonts/fonts';
+import '@/globals.css';
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+};
+
+export default async function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
   const { user, error, isAuthorized } = await getUser();
 
   if (error?.statusText === 'SESSION_EXPIRED' || !isAuthorized) redirect('/login');
 
   return (
-    <>
-      <Header user={user} />
-      {isAuthorized && error ? <ErrorTooltip /> : children}
-    </>
+    <html lang="en" suppressHydrationWarning>
+      <head />
+      <body className={`dark:bg-gray-900 ${geistSans.variable} ${geistMono.variable} antialiased`}>
+        <script
+          dangerouslySetInnerHTML={{
+            __html: `
+            (function() {
+              try {
+                const theme = localStorage.getItem('theme');
+                const defTheme = window.matchMedia("(prefers-color-scheme: dark)").matches;
+                if (theme === 'dark' || (!theme && defTheme)) {
+                  document.documentElement.setAttribute('data-theme', 'dark');
+                } else {
+                  document.documentElement.setAttribute('data-theme', 'light');
+                }
+              } catch (e) {}
+            })();
+          `,
+          }}
+        />
+        <div className="pr-3.5 pl-3.5">
+          <Header user={user} />
+          {isAuthorized && error ? <ErrorTooltip /> : children}
+        </div>
+      </body>
+    </html>
   );
 }
