+++ app/_shared/fonts/fonts.ts
@@ -1,0 +1,11 @@
+import { Geist, Geist_Mono } from 'next/font/google';
+
+export const geistSans = Geist({
+  variable: '--font-geist-sans',
+  subsets: ['latin'],
+});
+
+export const geistMono = Geist_Mono({
+  variable: '--font-geist-mono',
+  subsets: ['latin'],
+});
